requirements.txt
# ─────────────────────────────────────────────
#  Infrastructure
# ─────────────────────────────────────────────
boto3==1.34.0
python-dotenv==1.0.0

# ─────────────────────────────────────────────
#  Scraping (Bronze)
# ─────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.2
lxml

# ─────────────────────────────────────────────
#  ETL / Base de données (Silver & Gold)
# ─────────────────────────────────────────────
sqlalchemy
psycopg2-binary
pandas
numpy

# ─────────────────────────────────────────────
#  NLP & API
# ─────────────────────────────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
transformers==4.41.2
sentence-transformers==3.0.1
fastapi
uvicorn

# ─────────────────────────────────────────────
#  Airflow 2.9.3 — version stable
# ─────────────────────────────────────────────
apache-airflow==2.9.3
apache-airflow-providers-docker
apache-airflow-providers-postgres




loguru==0.7.2
tenacity==8.2.3




fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
boto3==1.34.0
pandas==2.2.2
pyarrow==16.0.0
pdfplumber==0.11.0
python-multipart==0.0.9
botocore==1.34.0
